Let users find the nar-maven-plugin website
There is a big "View details" button beneath the description of the NAR
plugin. Users are prone to click it without really reading the details in
the description, missing that the nar-maven-plugin moved to its own
website and is maintained there with great gratitude to the original
author.

Let's direct users interested in the current version of the NAR plugin to
the active website.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>